Fix: "cannot return value obtained from unexported field or method"
This error occurs when the struct has a private non-primitive type
like interface, slice, and etc. We use go:linkname to link the method
reflect.valueInterface(, value, false) and call it directly to bypass the
panic() (using the unsafe mode).

In addition, we add corresponding test cases for type complex and
interface().