Fix & Optimize: malloc of condExpr
1. Change the data type of `condExpr.sql` from string to *stringWriter. This reduce the
number of memory allocation if the sql is constructed through a concatenation.

2. Since the insert data is not fixed once set (Note: perviously, we use condExpr to store
the col insertion data), that means it internal data condExpr.sql is fixed during the life
cycle. In order to improve the performance, it can be optimized to `string`. Thus, we
create an internal data struct `valExpr` to store col insertion data.

3. In `stmt`, we add `whereRef []Cond` and `havingRef  []Cond` to track all `Cond`
created internally. They are freed once `stmt.Reset()` or `stmt.Destroy()` are called.
This fixes the memory leakage (`Cond` created internally, but not freed, thus collected
via GC).

4.  Move `Map.sortedKeys()` to sqlStmtMapUtilDebug.go. Thus, no more memory
allocation when calling catCondMap().