Fix the memory leakage when calling `Where()`, `And()`, or `Or()` with a string cond.
1. if the internal condition is not `condEmpty{}`, in the function `catCond()` case `string`,
`OpFunc()` construct either a `CondAnd` or `CondOr`. However, the created cond is
not saved in `*ref`, and will never get freed in `stmt.Reset()`. It relies on the GC to
collect. This commit detects if a new cond is created or not and saves it into `*ref`.

2. `OpFunc()` needs to create a temporary slice `[]Cond`. We optimize the allocation
with condAndPool as we did for `Map` and `Cond` cases.

3. make destroying the temporary slice `[]Cond` more readable via directly calling
`Destroy()`.