Feature: Add log, [level: debug, warning, info].
One can use `-tags=[level]` to determine the level of the log printed in the terminal. The `level` can
be `debug`, `warning`, and `info`. If `level` is not provided, only `error` messages are printed. Other
log function calls add zero overhead since these functions do nothing and are optimized by the
compiler.